Moved key linking function
